Codechange: Simplify scm file generation
This removes the external and hard-to-follow sed file, and instead uses
the shell `read` function to parse the xcf2png files and generate the
scm command from that (and changes the clean targets to follow the same
logic).

This makes the code a bit simpler, but also prepares for changing the
png output filenames from the Makefile in an upcoming commit.

This removes the case insensitive regex matching for for png/pcx
extensions, instead relying on the fact that all non-empty, non-comment
lines contain a image to generate (also, the rest of the code assumed
png and not pcx anyway it seems, or it was written to translate pcx to
png in xcf2png files, but that seems pointless now).

As a side effect, this might fix a bug that could happen if a png file
was generated but also mentioned in a comment (since the `$$j` file was
grepped twice, the second time not ignoring comments).

This commit should not change any output (with the current set of
xcf2png files), but does change the filenames of .scm files (which are
now .png.scm for simplicity).