Change: Add BUILD_DIR for generating intermediate files
Normally (and still by default), these are generated in (or under the
sprites subdir of) the repository root directory. For the png files,
this means that the pregenerated files from the sprites/png directory
are used (or updated).

If the (newly introduced) BUILD_DIR variable is modified, then all
intermediate files will be generated inside that directory. As a side
effect, all png files are regenerated from the xcf files.

This is convenient for the Debian packaging, which needs to regenerate
the png files from the xcf files (no prebuilt "binaries" are allowed).
By generating these files into a different directory, the original png
files are untouched and the source tree is unmodified after a build and
a clean.

The final bundle dir and archive files are still generated in the root
as before.